Lot's of changes
First commit in a while so there are a lot of changes. Some of the more important ones are that the file operations and ai code are locked behind features, but will compile for debug mode. Also a prelude module was added to lib.rs along with some functions, types, etc that didn't fit into one section or another.