Just a backup
I have done a lot of things in this one because it has been a while since I last put things on github. The most important ones are: creating a new file for some weird collection types I'm working on, deleting the dev readme because I clearly wasn't using it, and deleting partial iter ord because of how uterly broken it was.