dyn_binary and To/From Binary implementations
Made a new feature called dyn_binary meaning dynamic binary, which essentially is for stuff that does not have a constant size in a breaking way. For things like Vecs, while they do have a dynamic size, that is because we can put a number saying what size they are, but for something like a usize? It would be impractical to put a number before every single usize, isize and whatever other thing I've forgotten. So instead, by default, you cannot convert them to and from binary, and things that use them in their conversion by default (like Vecs because their length is stored as a usize) will instead us a 32 bit implementation. But if you do ever enable the feature, they will switch back to using their dynamically sized counter implementation