file_comm format 2: progress bar
Added in the second format(format #1) for file_comm, which has the main
focus of having a progress bar. Specifically, it has a progress bar, a
percent done, and a transfer rate average. Also made it so that it no
longer loops upon completion by default because when connecting to a
pre-specified server through the target argument, it would finish, then
do it again, and again, and again. Which is pointless because we already
did the thing.